Mention certificate file designed for Nginx
When I was playing around with SSL certificates, I accidentally opened the "Nginx Server" ZIP that was included in my certificates from WoSign (when they still offered free certificates) and StartSSL.  I was building my own certificate chain from the "Other Server" ZIP, which caused a few issues due to inexperience.  I think it would be helpful to mention that it should be an Nginx-specific file, but of course that's just me!